src/test/java/com/codingbat/array2/CountEvensTest.java
/*
  CodingBat Java Array 2 - https://codingbat.com/java/Array-2

  My repository - https://github.com/egalli64/jex/
 */
package com.codingbat.array2;

import static org.assertj.core.api.Assertions.assertThat;

import java.util.stream.Stream;

import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.MethodSource;

class CountEvensTest {
    static Stream<Arguments> provider() {
        return Stream.of( //
                Arguments.of(new int[] { 2, 1, 2, 3, 4 }, 3), //
                Arguments.of(new int[] { 2, 2, 0 }, 3), //
                Arguments.of(new int[] { 1, 3, 5 }, 0) //
        );
    }

    @ParameterizedTest
    @MethodSource("provider")
    void solutionExamples(int[] input, int expected) {
        assertThat(CountEvens.solution(input)).isEqualTo(expected);
    }
}
